// src/app/models/appearance.ts
export interface Appearance {
    lawFirmName: string;
    employerName: string;
    practiceArea: string;
    appearanceType: string;
    caseNo: string;
    address: string;
    recordName: string;
    courtName: string;
    appearanceDateTime: Date;
    minRate: number;
    maxRate: number;
    caseTitle: string;
    minExp: number;
    description: string;
    notes: string;
    status: string;
}